# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Quera Crypto Exchange server",
        "title": "QExchange-System",
        "contact": {
            "name": "Go-Zilla",
            "url": "https://github.com/Quera-Go-Zilla"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/block-user": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Blocks a user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Block user",
                "parameters": [
                    {
                        "description": "Block User",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.BlockUserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/admin/unblock-user": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Unblocks a user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Unblock user",
                "parameters": [
                    {
                        "description": "Unblock User",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UnblockUserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/admin/update-auth-level": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Updates a user's authentication level",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update user authentication level",
                "parameters": [
                    {
                        "description": "Update Auth Level",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateAuthenticationLevelRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/admin/update-to-admin": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Upgrades a regular user to an admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Upgrade user to admin",
                "parameters": [
                    {
                        "description": "Admin Password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateUserToAdminRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/admin/user-info": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Retrieves information for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get user information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.BlockUserRequest": {
            "type": "object",
            "properties": {
                "temporary": {
                    "type": "boolean",
                    "example": false
                },
                "username": {
                    "type": "string",
                    "example": "user1"
                }
            }
        },
...
        "handlers.UnblockUserRequest": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "example": "user2"
                }
            }
        },
        "handlers.UpdateAuthenticationLevelRequest": {
            "type": "object",
            "properties": {
                "new_auth_level": {
                    "type": "integer",
                    "example": 0
                },
                "username": {
                    "type": "string",
                    "example": "user2"
                }
            }
        },
        "handlers.UpdateUserToAdminRequest": {
            "type": "object",
            "properties": {
                "admin_password": {
                    "type": "string",
                    "example": "secret"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.UserInfo": {
            "type": "object",
            "properties": {
                "authentication_level": {
                    "description": "User's authentication level",
                    "type": "integer",
                    "example": 0
                },
                "balance": {
                    "description": "User's account balance",
                    "type": "integer",
                    "example": 5000000000
                },
                "banks_names": {
                    "description": "List of user's bank names",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['saman'",
                        " 'sepah']"
                    ]
                },
                "blocked_level": {
                    "description": "User's blocked level",
                    "type": "integer",
                    "example": 0
                },
                "closed_trades": {},
                "email": {
                    "description": "User's email address",
                    "type": "string",
                    "example": "user1@example.com"
                },
                "is_admin": {
                    "description": "Indicates if user has admin privileges",
                    "type": "boolean",
                    "example": false
                },
                "is_premium": {
                    "description": "Indicates if user has a premium account",
                    "type": "boolean",
                    "example": false
                },
                "open_trades": {},
                "phone_number": {
                    "description": "User's phone number",
                    "type": "string",
                    "example": "9876543210"
                },
                "username": {
                    "description": "User's username",
                    "type": "string",
                    "example": "user1"
                }
            }
        },
...
    }
}